+++ src/infrastructure/redis/src/test/resources/application-test.yml
@@ -1,0 +1,6 @@
+spring:
+    data:
+        redis:
+            host: ${REDIS_HOST}
+            port: ${REDIS_PORT}
+
