--- src/infrastructure/redis/src/testFixtures/kotlin/configuration/RedisTestConfiguration.kt
@@ -1,6 +1,0 @@
-package configuration
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication(scanBasePackages = ["yapp.be.redis"])
-class RedisTestConfiguration
