+++ src/infrastructure/redis/src/testFixtures/kotlin/container/RedisContainerProperties.kt
@@ -2,5 +2,5 @@ package container
 
 object RedisContainerProperties {
     const val imageName = "redis:alpine"
-    const val PORT = 6382
+    const val PORT = 6379
 }
