+++ src/infrastructure/storage/src/test/resources/application-test.yml
@@ -8,3 +8,7 @@ spring:
             hibernate:
                 format_sql: true
                 dialect: org.hibernate.dialect.MySQLDialect
+    datasource:
+        url: ${DATASOURCE_URL}
+        password: ${DATASOURCE_PASSWORD}
+        username: ${DATASOURCE_USERNAME}
